requirements.txt
python-telegram-bot==20.8
Flask==3.0.3
requests==2.32.2
